+++ Example/BitmovinAnalyticsCollector/AVFoundationViewController.swift
@@ -2,7 +2,7 @@ import AVKit
 import BitmovinAnalyticsCollector
 import UIKit
 
-class ViewController: UIViewController {
+class AVFoundationViewController: UIViewController {
     private static var playerViewControllerKVOContext = 0
     private var analyticsCollector: AVPlayerCollector
     private var isSeeking: Bool = false
@@ -40,7 +40,7 @@ class ViewController: UIViewController {
     }()
 
     required init?(coder aDecoder: NSCoder) {
-        config = BitmovinAnalyticsConfig(key: "9ae0b480-f2ee-4c10-bc3c-cb88e982e0ac")
+        config = BitmovinAnalyticsConfig(key: "e73a3577-d91c-4214-9e6d-938fb936818a")
         config.cdnProvider = CdnProvider.bitmovin
         config.customData1 = "customData1"
         config.customData2 = "customData2"
@@ -62,13 +62,13 @@ class ViewController: UIViewController {
 
     override func viewWillAppear(_: Bool) {
         playerView.playerLayer.player = player
-        addObserver(self, forKeyPath: #keyPath(ViewController.player.rate), options: [.new, .initial], context: &ViewController.playerViewControllerKVOContext)
+        addObserver(self, forKeyPath: #keyPath(AVFoundationViewController.player.rate), options: [.new, .initial], context: &AVFoundationViewController.playerViewControllerKVOContext)
         createPlayer()
         analyticsCollector.attachPlayer(player: player)
     }
 
     @IBAction func createPlayer() {
-        addObserver(self, forKeyPath: #keyPath(ViewController.player.currentItem.duration), options: [.new, .initial], context: &ViewController.playerViewControllerKVOContext)
+        addObserver(self, forKeyPath: #keyPath(AVFoundationViewController.player.currentItem.duration), options: [.new, .initial], context: &AVFoundationViewController.playerViewControllerKVOContext)
         let asset = AVURLAsset(url: url!, options: nil)
         player.replaceCurrentItem(with: AVPlayerItem(asset: asset))
         player.play()
@@ -84,12 +84,12 @@ class ViewController: UIViewController {
         analyticsCollector.detachPlayer()
         createPlayer()
         config.cdnProvider = CdnProvider.bitmovin
-        config.customData1 = "customData6"
-        config.customData2 = "customData7"
-        config.customData3 = "customData8"
-        config.customData4 = "customData9"
-        config.customData5 = "customData10"
-        config.customerUserId = "customUserId2"
+        config.customData1 = "customData1_2"
+        config.customData2 = "customData2_2"
+        config.customData3 = "customData3_2"
+        config.customData4 = "customData4_2"
+        config.customData5 = "customData5_2"
+        config.customerUserId = "customUserId_2"
         config.experimentName = "experiment-12"
         config.videoId = "iOSHLSStatic2"
         config.title = "ios static with AVFoundation"
@@ -99,12 +99,12 @@ class ViewController: UIViewController {
     }
 
     override func observeValue(forKeyPath keyPath: String?, of object: Any?, change: [NSKeyValueChangeKey: Any]?, context: UnsafeMutableRawPointer?) {
-        guard context == &ViewController.playerViewControllerKVOContext else {
+        guard context == &AVFoundationViewController.playerViewControllerKVOContext else {
             super.observeValue(forKeyPath: keyPath, of: object, change: change, context: context)
             return
         }
 
-        if keyPath == #keyPath(ViewController.player.currentItem.duration) {
+        if keyPath == #keyPath(AVPlayerViewController.player.currentItem.duration) {
             let newDuration: CMTime
             if let newDurationAsValue = change?[NSKeyValueChangeKey.newKey] as? NSValue {
                 newDuration = newDurationAsValue.timeValue
@@ -132,7 +132,7 @@ class ViewController: UIViewController {
 
             endDuration.isEnabled = hasValidDuration
             endDuration.text = createTimeString(time: Float(newDurationSeconds))
-        } else if keyPath == #keyPath(ViewController.player.rate) {
+        } else if keyPath == #keyPath(AVFoundationViewController.player.rate) {
             // Update `playPauseButton` image.
             let newRate = (change?[NSKeyValueChangeKey.newKey] as! NSNumber).doubleValue
             let buttonImageName = newRate > 0.0 ? "PauseButton" : "PlayButton"